Control the game with 2 shared controllers added playerscore to main Scene

+++ Assets/Game/Scripts/EventBus.cs
@@ -4,13 +4,13 @@ using UnityEngine;
 
 public class EventBus : MonoBehaviour
 {
-    public delegate void PlayerScore(PlayerController Player, float Score);
+    public delegate void PlayerScore(PlayerController Player, int Score);
     public static event PlayerScore OnPlayerScored;
 
     void Start() { }
     void Update() { }
 
-    public static void PlayerScored(PlayerController Player, float Score)
+    public static void PlayerScored(PlayerController Player, int Score)
     {
         if (OnPlayerScored != null) { OnPlayerScored(Player, Score); }
     }
